+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 2026-07-30
+
 ### Changed
 
 - Widened `OrderedCollections` compat to `"1, 2"`. This was previously untestable —